fix(constants file): Rename all constants file to {Component_Name}Constants
