+++ src/components/Alert/helpers.test.js
@@ -1,5 +1,5 @@
 import { warnIfDeprecatedType, getIconName, getClassName } from './helpers';
-import { ALERT_TYPES, DEPRECATED_ALERT_TYPES } from './constants';
+import { ALERT_TYPES, DEPRECATED_ALERT_TYPES } from './AlertConstants';
 
 const ALL_ALERT_TYPES = [...ALERT_TYPES, ...DEPRECATED_ALERT_TYPES];
