+++ src/components/Button/DropdownButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { DropdownButton as BsDropdownButton } from 'react-bootstrap';
-import { BUTTON_BS_STYLES } from './constants';
+import { BUTTON_BS_STYLES } from './ButtonConstants';
 
 const DropdownButton = props => <BsDropdownButton {...props} />;
