+++ src/components/Chart/AreaChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import C3Chart from 'react-c3js';
-import { getComposer } from './constants';
+import { getComposer } from './ChartConstants';
 
 const AreaChart = getComposer('AREA_CHART')(
   ({ className, type, data, ...props }) => (
