+++ src/components/Chart/GroupedBarChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import C3Chart from 'react-c3js';
-import { getComposer } from './constants';
+import { getComposer } from './ChartConstants';
 
 const GroupedBarChart = getComposer('GROUPED_BAR_CHART')(
   ({ className, type, data, ...props }) => (
