+++ src/components/Chart/SingleLineChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import C3Chart from 'react-c3js';
-import { getComposer } from './constants';
+import { getComposer } from './ChartConstants';
 
 const SingleLineChart = getComposer('SINGLE_LINE_CHART')(
   ({ className, type, data, ...props }) => (
