+++ src/components/Chart/SparklineChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import C3Chart from 'react-c3js';
-import { getComposer } from './constants';
+import { getComposer } from './ChartConstants';
 
 const SparklineChart = getComposer('SPARKLINE_CHART')(
   ({ className, type, data, ...props }) => (
