+++ src/components/Pagination/PaginationConstants.js
